+++ integrated_channels/sap_success_factors/migrations/0008_historicalsapsuccessfactorsenterprisecustomerconfiguration_history_change_reason.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('sap_success_factors', '0007_remove_historicalsapsuccessfactorsenterprisecustomerconfiguration_history_change_reason'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='historicalsapsuccessfactorsenterprisecustomerconfiguration',
+            name='history_change_reason',
+            field=models.CharField(max_length=100, null=True),
+        ),
+    ]